+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 
 This document records the main changes to the ``sdssdb`` code.
 
+* Add ``ForeignKeyField`` from ``CatalogToSDSS_DR13_PhotoObj_Primary`` directly to ``SDSS_DR13_PhotoObj``.
+
 * :release:`0.4.4 <2020-07-07>`
 * Add ``ForeignKeyField`` to ``TIC_v8`` where for all models in ``catalogdb`` connected to Gaia.
 * Use ``autorollback=True`` by default in `.PeeweeDatabaseConnection`.